removed bottomNavigation.dart + moved and renamed tabs.dart to footer\bottomNavigationTabs.dart
