--- lib/components/footer/bottomNavigation.dart
@@ -1,32 +1,0 @@
-import 'package:flutter/material.dart';
-
-@Deprecated('removed bottomNavigation, this is now un-needed')
-final footer = new BottomNavigationBar(
-    showSelectedLabels: true,
-    backgroundColor: Colors.deepPurple,
-    items: [
-      new BottomNavigationBarItem(
-          icon: Icon(
-            Icons.home,
-            color: Colors.white,
-            size: 30.0,
-          ),
-          title: Text('Home'),
-          backgroundColor: Colors.white),
-      new BottomNavigationBarItem(
-          icon: Icon(
-            Icons.receipt,
-            color: Colors.white,
-            size: 30.0,
-          ),
-          title: Text('Reports'),
-          backgroundColor: Colors.white),
-      new BottomNavigationBarItem(
-          icon: Icon(
-            Icons.settings,
-            color: Colors.white,
-            size: 30.0,
-          ),
-          title: Text('Settings'),
-          backgroundColor: Colors.white),
-    ]);
